+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 2.26.0 (IN PROGRESS)
 
+* Implement permission assigment by batch. Refs UIU-1249
+
 ## [2.25.3](https://github.com/folio-org/ui-users/tree/v2.25.3) (2019-09-26)
 [Full Changelog](https://github.com/folio-org/ui-users/compare/v2.25.2...v2.25.3)